requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
edge-tts>=7.0.0
groq>=0.4.0
feedparser>=6.0.0
requests>=2.31.0
pydub>=0.25.0
mutagen>=1.47.0
